+++ EPlast/EPlast.BLL/Interfaces/Auth/IAuthService.cs
@@ -1,42 +1,22 @@
 ﻿using EPlast.BLL.DTO.Account;
 using EPlast.BLL.DTO.UserProfiles;
+using EPlast.BLL.Models;
+using EPlast.DataAccess.Entities;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Identity;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using EPlast.BLL.Models;
-using EPlast.DataAccess.Entities;
 
 namespace EPlast.BLL.Interfaces
 {
     public interface IAuthService
     {
         /// <summary>
-        /// Login in system
-        /// </summary>
-        /// <param name="loginDto"></param>
-        /// <returns>Result of logining in system</returns>
-        Task<SignInResult> SignInAsync(LoginDto loginDto);
-
-        /// <summary>
-        /// Logout in system
-        /// </summary>
-        void SignOutAsync();
-
-        /// <summary>
-        /// Creating user in database
+        /// Add role and token for user after registration
         /// </summary>
         /// <param name="registerDto"></param>
-        /// <returns>Result of creating user in system</returns>
-        Task<IdentityResult> CreateUserAsync(RegisterDto registerDto);
-
-        /// <summary>
-        /// Confirming Email after registration
-        /// </summary>
-        /// <param name="userId"></param>
-        /// <param name="code"></param>
-        /// <returns>Result of confirming email in system</returns>
-        Task<IdentityResult> ConfirmEmailAsync(string userId, string code);
+        /// <returns>Result of adding role and token</returns>
+        Task<string> AddRoleAndTokenAsync(string email);
 
         /// <summary>
         /// Changing password for user account
@@ -47,69 +27,6 @@ namespace EPlast.BLL.Interfaces
         Task<IdentityResult> ChangePasswordAsync(string userId, ChangePasswordDto changePasswordDto);
 
         /// <summary>
-        /// Refresh signin credentials
-        /// </summary>
-        /// <param name="userDto"></param>
-        void RefreshSignInAsync(UserDTO userDto);
-
-        /// <summary>
-        /// Get authentication properties for user
-        /// </summary>
-        /// <param name="provider"></param>
-        /// <param name="returnUrl"></param>
-        /// <returns>Authentication properties for user</returns>
-        AuthenticationProperties GetAuthProperties(string provider, string returnUrl);
-
-        /// <summary>
-        /// Get information about external login
-        /// </summary>
-        /// <returns>External login information</returns>
-        Task<ExternalLoginInfo> GetInfoAsync();
-
-        /// <summary>
-        /// Get sign in result
-        /// </summary>
-        /// <param name="externalLoginInfo"></param>
-        /// <returns>Sign in result</returns>
-        Task<SignInResult> GetSignInResultAsync(ExternalLoginInfo externalLoginInfo);
-
-        /// <summary>
-        /// Checking if email was confirmed
-        /// </summary>
-        /// <param name="userDto"></param>
-        /// <returns>Result of confirming email</returns>
-        Task<bool> IsEmailConfirmedAsync(UserDTO userDto);
-
-        /// <summary>
-        /// Add role and token for user after registration
-        /// </summary>
-        /// <param name="registerDto"></param>
-        /// <returns>Result of adding role and token</returns>
-        Task<string> AddRoleAndTokenAsync(string email);
-
-        /// <summary>
-        /// Generating confirmation token
-        /// </summary>
-        /// <param name="userDto"></param>
-        /// <returns>Result of generating token</returns>
-        Task<string> GenerateConfToken(UserDTO userDto);
-
-        /// <summary>
-        /// Generating reset token 
-        /// </summary>
-        /// <param name="userDto"></param>
-        /// <returns>Result of generating reset token</returns>
-        Task<string> GenerateResetTokenAsync(UserDTO userDto);
-
-        /// <summary>
-        /// Reseting password for user
-        /// </summary>
-        /// <param name="userId"></param>
-        /// <param name="resetPasswordDto"></param>
-        /// <returns>Returs result of reseting password</returns>
-        Task<IdentityResult> ResetPasswordAsync(string userId, ResetPasswordDto resetPasswordDto);
-
-        /// <summary>
         /// Checking if user was locked
         /// </summary>
         /// <param name="userDto"></param>
@@ -117,10 +34,17 @@ namespace EPlast.BLL.Interfaces
         Task CheckingForLocking(UserDTO userDto);
 
         /// <summary>
-        /// Get authentication scheme for user
+        /// Creating user in database
         /// </summary>
-        /// <returns>Returns authentication scheme</returns>
-        Task<IEnumerable<AuthenticationScheme>> GetAuthSchemesAsync();
+        /// <param name="registerDto"></param>
+        /// <returns>Result of creating user in system</returns>
+        Task<IdentityResult> CreateUserAsync(RegisterDto registerDto);
+
+        /// <summary>
+        /// Sign in using Facebook
+        /// </summary>
+        /// <param name="facebookUser"></param>
+        Task<UserDTO> FacebookLoginAsync(FacebookUserInfo facebookUser);
 
         /// <summary>
         /// Finding by email in database
@@ -137,11 +61,59 @@ namespace EPlast.BLL.Interfaces
         Task<UserDTO> FindByIdAsync(string id);
 
         /// <summary>
+        /// Generating confirmation token
+        /// </summary>
+        /// <param name="userDto"></param>
+        /// <returns>Result of generating token</returns>
+        Task<string> GenerateConfToken(UserDTO userDto);
+
+        /// <summary>
+        /// Generating reset token
+        /// </summary>
+        /// <param name="userDto"></param>
+        /// <returns>Result of generating reset token</returns>
+        Task<string> GenerateResetTokenAsync(UserDTO userDto);
+
+        /// <summary>
+        /// Get authentication properties for user
+        /// </summary>
+        /// <param name="provider"></param>
+        /// <param name="returnUrl"></param>
+        /// <returns>Authentication properties for user</returns>
+        AuthenticationProperties GetAuthProperties(string provider, string returnUrl);
+
+        /// <summary>
+        /// Get authentication scheme for user
+        /// </summary>
+        /// <returns>Returns authentication scheme</returns>
+        Task<IEnumerable<AuthenticationScheme>> GetAuthSchemesAsync();
+
+        /// <summary>
+        /// Validates google token and gets user information
+        /// </summary>
+        /// <param name="providerToken"></param>
+        /// <returns>Returns Google user information</returns>
+        Task<UserDTO> GetGoogleUserAsync(string providerToken);
+
+        /// <summary>
         /// Get id for user from database
         /// </summary>
         /// <param name="user"></param>
         /// <returns>Returns id of user</returns>
         Task<string> GetIdForUser(User user);
+
+        /// <summary>
+        /// Get information about external login
+        /// </summary>
+        /// <returns>External login information</returns>
+        Task<ExternalLoginInfo> GetInfoAsync();
+
+        /// <summary>
+        /// Get sign in result
+        /// </summary>
+        /// <param name="externalLoginInfo"></param>
+        /// <returns>Sign in result</returns>
+        Task<SignInResult> GetSignInResultAsync(ExternalLoginInfo externalLoginInfo);
 
         /// <summary>
         /// Returns time after registration for user
@@ -165,16 +137,36 @@ namespace EPlast.BLL.Interfaces
         UserDTO GetUser(User user);
 
         /// <summary>
-        ///  Validates google token and gets user information
+        /// Checking if email was confirmed
         /// </summary>
-        /// <param name="providerToken"></param>
-        /// <returns>Returns Google user information</returns>
-        Task<UserDTO> GetGoogleUserAsync(string providerToken);
+        /// <param name="userDto"></param>
+        /// <returns>Result of confirming email</returns>
+        Task<bool> IsEmailConfirmedAsync(UserDTO userDto);
 
         /// <summary>
-        /// Sign in using Facebook
+        /// Refresh signin credentials
         /// </summary>
-        /// <param name="facebookUser"></param>
-        Task<UserDTO> FacebookLoginAsync(FacebookUserInfo facebookUser);
+        /// <param name="userDto"></param>
+        void RefreshSignInAsync(UserDTO userDto);
+
+        /// <summary>
+        /// Reseting password for user
+        /// </summary>
+        /// <param name="userId"></param>
+        /// <param name="resetPasswordDto"></param>
+        /// <returns>Returs result of reseting password</returns>
+        Task<IdentityResult> ResetPasswordAsync(string userId, ResetPasswordDto resetPasswordDto);
+
+        /// <summary>
+        /// Login in system
+        /// </summary>
+        /// <param name="loginDto"></param>
+        /// <returns>Result of logining in system</returns>
+        Task<SignInResult> SignInAsync(LoginDto loginDto);
+
+        /// <summary>
+        /// Logout in system
+        /// </summary>
+        void SignOutAsync();
     }
 }
