+++ EPlast/EPlast.BLL/Services/Auth/AuthService.cs
@@ -2,9 +2,13 @@
 using EPlast.BLL.DTO.Account;
 using EPlast.BLL.DTO.UserProfiles;
 using EPlast.BLL.Interfaces;
+using EPlast.BLL.Models;
 using EPlast.DataAccess.Entities;
+using EPlast.DataAccess.Repositories;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Identity;
+using Newtonsoft.Json;
+using NLog.Extensions.Logging;
 using System;
 using System.Collections.Generic;
 using System.Globalization;
@@ -12,20 +16,16 @@ using System.Linq;
 using System.Net;
 using System.Net.Http;
 using System.Threading.Tasks;
-using EPlast.BLL.Models;
-using EPlast.DataAccess.Repositories;
-using Newtonsoft.Json;
-using NLog.Extensions.Logging;
 
 namespace EPlast.BLL.Services
 {
     public class AuthService : IAuthService
     {
-        private readonly UserManager<User> _userManager;
-        private readonly SignInManager<User> _signInManager;
         private readonly IEmailConfirmation _emailConfirmation;
         private readonly IMapper _mapper;
         private readonly IRepositoryWrapper _repoWrapper;
+        private readonly SignInManager<User> _signInManager;
+        private readonly UserManager<User> _userManager;
         public AuthService(
             UserManager<User> userManager,
             SignInManager<User> signInManager,
@@ -41,17 +41,30 @@ namespace EPlast.BLL.Services
         }
 
         ///<inheritdoc/>
-        public async Task<SignInResult> SignInAsync(LoginDto loginDto)
+        public async Task<string> AddRoleAndTokenAsync(string email)
         {
-            var user = _userManager.FindByEmailAsync(loginDto.Email);
-            var result = await _signInManager.PasswordSignInAsync(user.Result, loginDto.Password, loginDto.RememberMe, true);
+            var user = await _userManager.FindByEmailAsync(email);
+            await _userManager.AddToRoleAsync(user, "Зареєстрований користувач");
+            string code = await _userManager.GenerateEmailConfirmationTokenAsync(user);
+            return code;
+        }
+
+        ///<inheritdoc/>
+        public async Task<IdentityResult> ChangePasswordAsync(string userId, ChangePasswordDto changePasswordDto)
+        {
+            var user = await _userManager.FindByIdAsync(userId);
+            var result = await _userManager.ChangePasswordAsync(user, changePasswordDto.CurrentPassword,
+                changePasswordDto.NewPassword);
             return result;
         }
 
         ///<inheritdoc/>
-        public async void SignOutAsync()
+        public async Task CheckingForLocking(UserDTO userDto)
         {
-            await _signInManager.SignOutAsync();
+            if (await _userManager.IsLockedOutAsync(_mapper.Map<UserDTO, User>(userDto)))
+            {
+                await _userManager.SetLockoutEndDateAsync(_mapper.Map<UserDTO, User>(userDto), DateTimeOffset.UtcNow);
+            }
         }
 
         ///<inheritdoc/>
@@ -73,65 +86,51 @@ namespace EPlast.BLL.Services
             return result;
         }
 
-        ///<inheritdoc/>
-        public async Task<IdentityResult> ConfirmEmailAsync(string userId, string code)
+        public async Task<UserDTO> FacebookLoginAsync(FacebookUserInfo facebookUser)
         {
-            var user = await _userManager.FindByIdAsync(userId);
-            var result = await _userManager.ConfirmEmailAsync(user, code);
-            return result;
+            var user = await _userManager.FindByEmailAsync(facebookUser.Email);
+            if (user == null)
+            {
+                user = new User
+                {
+                    SocialNetworking = true,
+                    UserName = facebookUser.Email ?? facebookUser.UserId,
+                    FirstName = facebookUser.Name.Split(' ')[0],
+                    Email = facebookUser.Email ?? "facebookdefaultmail@gmail.com",
+                    LastName = facebookUser.Name.Split(' ')[1],
+                    ImagePath = "default_user_image.png",
+                    EmailConfirmed = true,
+                    RegistredOn = DateTime.Now,
+                    UserProfile = new UserProfile()
+                    {
+                        Birthday = DateTime.Parse(facebookUser.Birthday, CultureInfo.InvariantCulture),
+                        GenderID = _repoWrapper.Gender.FindByCondition(x => x.Name == facebookUser.Gender).FirstOrDefault()?.ID,
+                    }
+                };
+                var createResult = await _userManager.CreateAsync(user);
+                if (createResult.Succeeded && user.Email != "facebookdefaultmail@gmail.com")
+                {
+                    await _emailConfirmation.SendEmailAsync(user.Email, "Повідомлення про реєстрацію",
+                        "Ви зареєструвались в системі EPlast використовуючи свій Facebook-акаунт. ", "Адміністрація сайту EPlast");
+                }
+                await _userManager.AddToRoleAsync(user, "Прихильник");
+            }
+            await _signInManager.SignInAsync(user, isPersistent: false);
+            return _mapper.Map<User, UserDTO>(user);
         }
 
         ///<inheritdoc/>
-        public async Task<IdentityResult> ChangePasswordAsync(string userId, ChangePasswordDto changePasswordDto)
-        {
-            var user = await _userManager.FindByIdAsync(userId);
-            var result = await _userManager.ChangePasswordAsync(user, changePasswordDto.CurrentPassword,
-                changePasswordDto.NewPassword);
-            return result;
-        }
-
-        ///<inheritdoc/>
-        public async void RefreshSignInAsync(UserDTO userDto)
-        {
-            await _signInManager.RefreshSignInAsync(_mapper.Map<UserDTO, User>(userDto));
-        }
-
-        ///<inheritdoc/>
-        public AuthenticationProperties GetAuthProperties(string provider, string returnUrl)
-        {
-            AuthenticationProperties properties = _signInManager.ConfigureExternalAuthenticationProperties(provider, returnUrl);
-            return properties;
-        }
-
-        ///<inheritdoc/>
-        public async Task<ExternalLoginInfo> GetInfoAsync()
-        {
-            var info = await _signInManager.GetExternalLoginInfoAsync();
-            return info;
-        }
-
-        ///<inheritdoc/>
-        public async Task<SignInResult> GetSignInResultAsync(ExternalLoginInfo externalLoginInfo)
-        {
-            SignInResult signInResult = await _signInManager.ExternalLoginSignInAsync(externalLoginInfo.LoginProvider,
-                    externalLoginInfo.ProviderKey, isPersistent: false, bypassTwoFactor: true);
-            return signInResult;
-        }
-
-        ///<inheritdoc/>
-        public async Task<bool> IsEmailConfirmedAsync(UserDTO userDto)
-        {
-            bool result = await _userManager.IsEmailConfirmedAsync(_mapper.Map<UserDTO, User>(userDto));
-            return result;
-        }
-
-        ///<inheritdoc/>
-        public async Task<string> AddRoleAndTokenAsync(string email)
+        public async Task<UserDTO> FindByEmailAsync(string email)
         {
             var user = await _userManager.FindByEmailAsync(email);
-            await _userManager.AddToRoleAsync(user, "Зареєстрований користувач");
-            string code = await _userManager.GenerateEmailConfirmationTokenAsync(user);
-            return code;
+            return _mapper.Map<User, UserDTO>(user);
+        }
+
+        ///<inheritdoc/>
+        public async Task<UserDTO> FindByIdAsync(string id)
+        {
+            var user = await _userManager.FindByIdAsync(id);
+            return _mapper.Map<User, UserDTO>(user);
         }
 
         ///<inheritdoc/>
@@ -150,20 +149,10 @@ namespace EPlast.BLL.Services
         }
 
         ///<inheritdoc/>
-        public async Task<IdentityResult> ResetPasswordAsync(string userId, ResetPasswordDto resetPasswordDto)
+        public AuthenticationProperties GetAuthProperties(string provider, string returnUrl)
         {
-            var user = await _userManager.FindByIdAsync(userId);
-            var result = await _userManager.ResetPasswordAsync(user, resetPasswordDto.Code, resetPasswordDto.Password);
-            return result;
-        }
-
-        ///<inheritdoc/>
-        public async Task CheckingForLocking(UserDTO userDto)
-        {
-            if (await _userManager.IsLockedOutAsync(_mapper.Map<UserDTO, User>(userDto)))
-            {
-                await _userManager.SetLockoutEndDateAsync(_mapper.Map<UserDTO, User>(userDto), DateTimeOffset.UtcNow);
-            }
+            AuthenticationProperties properties = _signInManager.ConfigureExternalAuthenticationProperties(provider, returnUrl);
+            return properties;
         }
 
         ///<inheritdoc/>
@@ -171,51 +160,6 @@ namespace EPlast.BLL.Services
         {
             var externalLogins = await _signInManager.GetExternalAuthenticationSchemesAsync();
             return externalLogins;
-        }
-
-        ///<inheritdoc/>
-        public async Task<UserDTO> FindByEmailAsync(string email)
-        {
-            var user = await _userManager.FindByEmailAsync(email);
-            return _mapper.Map<User, UserDTO>(user);
-        }
-
-        ///<inheritdoc/>
-        public async Task<UserDTO> FindByIdAsync(string id)
-        {
-            var user = await _userManager.FindByIdAsync(id);
-            return _mapper.Map<User, UserDTO>(user);
-        }
-
-        ///<inheritdoc/>
-        public async Task<string> GetIdForUser(User user)
-        {
-            var currentUserId = await _userManager.GetUserIdAsync(user);
-            return currentUserId;
-        }
-
-        ///<inheritdoc/>
-        public int GetTimeAfterRegistr(UserDTO userDto)
-        {
-            IDateTimeHelper dateTimeConfirming = new DateTimeHelper();
-            var user = _mapper.Map<UserDTO, User>(userDto);
-            int totalTime = (int)dateTimeConfirming.GetCurrentTime().Subtract(user.EmailSendedOnRegister).TotalMinutes;
-            return totalTime;
-        }
-
-        ///<inheritdoc/>
-        public int GetTimeAfterReset(UserDTO userDto)
-        {
-            IDateTimeHelper dateTimeResetingPassword = new DateTimeHelper();
-            dateTimeResetingPassword.GetCurrentTime();
-            int totalTime = (int)dateTimeResetingPassword.GetCurrentTime().Subtract(userDto.EmailSendedOnForgotPassword).TotalMinutes;
-            return totalTime;
-        }
-
-        ///<inheritdoc/>
-        public UserDTO GetUser(User user)
-        {
-            return _mapper.Map<User, UserDTO>(user);
         }
 
         ///<inheritdoc/>
@@ -262,39 +206,87 @@ namespace EPlast.BLL.Services
             }
             await _signInManager.SignInAsync(user, isPersistent: false);
             return _mapper.Map<User, UserDTO>(user);
-
         }
-        public async Task<UserDTO> FacebookLoginAsync(FacebookUserInfo facebookUser)
+
+        ///<inheritdoc/>
+        public async Task<string> GetIdForUser(User user)
         {
-            var user = await _userManager.FindByEmailAsync(facebookUser.Email);
-            if (user == null)
-            {
-                user = new User
-                {
-                    SocialNetworking = true,
-                    UserName = facebookUser.Email ?? facebookUser.UserId,
-                    FirstName = facebookUser.Name.Split(' ')[0],
-                    Email = facebookUser.Email ?? "facebookdefaultmail@gmail.com",
-                    LastName = facebookUser.Name.Split(' ')[1],
-                    ImagePath = "default_user_image.png",
-                    EmailConfirmed = true,
-                    RegistredOn = DateTime.Now,
-                    UserProfile = new UserProfile()
-                    {
-                        Birthday = DateTime.Parse(facebookUser.Birthday, CultureInfo.InvariantCulture),
-                        GenderID = _repoWrapper.Gender.FindByCondition(x => x.Name == facebookUser.Gender).FirstOrDefault()?.ID,
-                    }
-                };
-                var createResult = await _userManager.CreateAsync(user);
-                if (createResult.Succeeded && user.Email != "facebookdefaultmail@gmail.com")
-                {
-                    await _emailConfirmation.SendEmailAsync(user.Email, "Повідомлення про реєстрацію",
-                        "Ви зареєструвались в системі EPlast використовуючи свій Facebook-акаунт. ", "Адміністрація сайту EPlast");
-                }
-                await _userManager.AddToRoleAsync(user, "Прихильник");
-            }
-            await _signInManager.SignInAsync(user, isPersistent: false);
+            var currentUserId = await _userManager.GetUserIdAsync(user);
+            return currentUserId;
+        }
+
+        ///<inheritdoc/>
+        public async Task<ExternalLoginInfo> GetInfoAsync()
+        {
+            var info = await _signInManager.GetExternalLoginInfoAsync();
+            return info;
+        }
+
+        ///<inheritdoc/>
+        public async Task<SignInResult> GetSignInResultAsync(ExternalLoginInfo externalLoginInfo)
+        {
+            SignInResult signInResult = await _signInManager.ExternalLoginSignInAsync(externalLoginInfo.LoginProvider,
+                    externalLoginInfo.ProviderKey, isPersistent: false, bypassTwoFactor: true);
+            return signInResult;
+        }
+
+        ///<inheritdoc/>
+        public int GetTimeAfterRegistr(UserDTO userDto)
+        {
+            IDateTimeHelper dateTimeConfirming = new DateTimeHelper();
+            var user = _mapper.Map<UserDTO, User>(userDto);
+            int totalTime = (int)dateTimeConfirming.GetCurrentTime().Subtract(user.EmailSendedOnRegister).TotalMinutes;
+            return totalTime;
+        }
+
+        ///<inheritdoc/>
+        public int GetTimeAfterReset(UserDTO userDto)
+        {
+            IDateTimeHelper dateTimeResetingPassword = new DateTimeHelper();
+            dateTimeResetingPassword.GetCurrentTime();
+            int totalTime = (int)dateTimeResetingPassword.GetCurrentTime().Subtract(userDto.EmailSendedOnForgotPassword).TotalMinutes;
+            return totalTime;
+        }
+
+        ///<inheritdoc/>
+        public UserDTO GetUser(User user)
+        {
             return _mapper.Map<User, UserDTO>(user);
+        }
+
+        ///<inheritdoc/>
+        public async Task<bool> IsEmailConfirmedAsync(UserDTO userDto)
+        {
+            bool result = await _userManager.IsEmailConfirmedAsync(_mapper.Map<UserDTO, User>(userDto));
+            return result;
+        }
+
+        ///<inheritdoc/>
+        public async void RefreshSignInAsync(UserDTO userDto)
+        {
+            await _signInManager.RefreshSignInAsync(_mapper.Map<UserDTO, User>(userDto));
+        }
+
+        ///<inheritdoc/>
+        public async Task<IdentityResult> ResetPasswordAsync(string userId, ResetPasswordDto resetPasswordDto)
+        {
+            var user = await _userManager.FindByIdAsync(userId);
+            var result = await _userManager.ResetPasswordAsync(user, resetPasswordDto.Code, resetPasswordDto.Password);
+            return result;
+        }
+
+        ///<inheritdoc/>
+        public async Task<SignInResult> SignInAsync(LoginDto loginDto)
+        {
+            var user = _userManager.FindByEmailAsync(loginDto.Email);
+            var result = await _signInManager.PasswordSignInAsync(user.Result, loginDto.Password, loginDto.RememberMe, true);
+            return result;
+        }
+
+        ///<inheritdoc/>
+        public async void SignOutAsync()
+        {
+            await _signInManager.SignOutAsync();
         }
     }
 }
